pyproject.toml
[tool.poetry]
name = "imagesearch"
version = "0.1.0"
description = ""
authors = ["Upa Gupta <upagupta2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
python-multipart = "^0.0.17"
pillow = "^11.0.0"
requests = "^2.32.3"
numpy = "^2.1.3"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
google-api-python-client = "^2.151.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
chromadb = "^0.5.18"
pysqlite3-binary = "^0.5.4"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.3"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
